// src/utils/interfaces/articles/get-all-articles.types.ts
export interface GetAllArticlesResponse {
	id: { type: string; value: string };
	title: { type: string; value: string };
	authors: { type: string; value: string };
	year: { type: string; value: string };
	publisher: { type: string; value: string };
	abstract: { type: string; value: string };
	doi: { type: string; value: string };
	keywords: { type: string; value: string };
	url: { type: string; value: string };
	authorsId: { type: string; value: string };
}